Remove model gradients tests (replaced with hessian info calculator tests in both frameworks)
