+++ tests/pytorch_tests/function_tests/test_function_runner.py
@@ -24,16 +24,15 @@ from tests.pytorch_tests.function_tests.kpi_data_test import TestKPIDataBasicAll
     TestKPIDataBasicPartialBitwidth, TestKPIDataComplexPartialBitwidth, TestKPIDataComplesAllBitwidth
 from tests.pytorch_tests.function_tests.layer_fusing_test import LayerFusingTest1, LayerFusingTest2, LayerFusingTest3, \
     LayerFusingTest4
-from tests.pytorch_tests.function_tests.model_gradients_test import ModelGradientsBasicModelTest, \
-    ModelGradientsCalculationTest, ModelGradientsAdvancedModelTest, \
-    ModelGradientsMultipleOutputsModelTest, ModelGradientsNonDifferentiableNodeModelTest, \
-    ModelGradientsMultipleOutputsTest, ModelGradientsSinglePointTest
 from tests.pytorch_tests.function_tests.set_layer_to_bitwidth_test import TestSetLayerToBitwidthWeights, \
     TestSetLayerToBitwidthActivation
 from tests.pytorch_tests.function_tests.test_sensitivity_eval_non_supported_output import \
     TestSensitivityEvalWithArgmaxNode
-from tests.pytorch_tests.function_tests.test_hessian_info_weights import WeightsHessianTraceBasicModelTest, WeightsHessianTraceAdvanceModelTest, \
-WeightsHessianTraceMultipleOutputsModelTest, WeightsHessianTraceReuseModelTest
+from tests.pytorch_tests.function_tests.test_hessian_info_calculator import WeightsHessianTraceBasicModelTest, \
+    WeightsHessianTraceAdvanceModelTest, \
+    WeightsHessianTraceMultipleOutputsModelTest, WeightsHessianTraceReuseModelTest, \
+    ActivationHessianTraceBasicModelTest, ActivationHessianTraceAdvanceModelTest, \
+    ActivationHessianTraceMultipleOutputsModelTest, ActivationHessianTraceReuseModelTest
 
 
 class FunctionTestRunner(unittest.TestCase):
@@ -101,21 +100,18 @@ class FunctionTestRunner(unittest.TestCase):
         """
         TestKPIDataComplexPartialBitwidth(self).run_test()
 
-    def test_model_gradients(self):
+    def test_activation_hessian_trace(self):
         """
-        This test checks the Model Gradients Pytorch computation.
+        This test checks the activation hessian trace approximation in Pytorch.
         """
-        ModelGradientsBasicModelTest(self).run_test()
-        ModelGradientsCalculationTest(self).run_test()
-        ModelGradientsAdvancedModelTest(self).run_test()
-        ModelGradientsMultipleOutputsTest(self).run_test()
-        ModelGradientsMultipleOutputsModelTest(self).run_test()
-        ModelGradientsNonDifferentiableNodeModelTest(self).run_test()
-        ModelGradientsSinglePointTest(self).run_test()
+        ActivationHessianTraceBasicModelTest(self).run_test()
+        ActivationHessianTraceAdvanceModelTest(self).run_test()
+        ActivationHessianTraceMultipleOutputsModelTest(self).run_test()
+        ActivationHessianTraceReuseModelTest(self).run_test()
 
     def test_weights_hessian_trace(self):
         """
-        This test checks the weighes hessian trace approximation in Pytorch.
+        This test checks the weights hessian trace approximation in Pytorch.
         """
         WeightsHessianTraceBasicModelTest(self).run_test()
         WeightsHessianTraceAdvanceModelTest(self).run_test()
@@ -141,7 +137,7 @@ class FunctionTestRunner(unittest.TestCase):
 
     def test_sensitivity_eval_not_supported_output(self):
         """
-        This test verifies failure on non supported output nodes in mixed precision with Hessian-based scores.
+        This test verifies failure on non-supported output nodes in mixed precision with Hessian-based scores.
         """
         TestSensitivityEvalWithArgmaxNode(self).run_test()
